requirements.txt
fastapi
uvicorn
pydantic
azure-ai-formrecognizer
azure-core
azure-storage-blob
azure-data-tables
openai
pyotp
qrcode
bcrypt
python-dotenv
azure-search-documents